[B3] Add file app_convert to project
