+++ esp_ble/main/CMakeLists.txt
@@ -1,2 +1,2 @@
-idf_component_register(SRCS "app_main.c" "app_uart.c" "app_ble.c" "app_flash.c" "common_ble.c" "app_timer.c"
+idf_component_register(SRCS "app_convert.c" "app_main.c" "app_uart.c" "app_ble.c" "app_flash.c" "common_ble.c" "app_timer.c"
                     INCLUDE_DIRS ".")
