+++ esp_ble/main/app_convert.h
@@ -1,0 +1,7 @@
+#ifndef _APP_CONVERT_H_
+#define _APP_CONVERT_H_
+
+/* Convert char to dec */
+uint32_t app_convert_char2Dec(uint8_t ch1, uint8_t ch2);
+
+#endif /* _APP_CONVERT_H_ */
