+++ esp_ble/main/app_main.c
@@ -11,6 +11,7 @@
 #include "app_uart.h"
 #include "esp_log.h"
 #include "app_timer.h"
+#include "app_convert.h"
 
 static const char *TAG = "MAIN_APP";
 static uint8_t index_for_connected_to_peer = 0;
@@ -30,8 +31,6 @@ static void app_parse_data_from_uart(uint8_t data[], uint32_t length);
 void app_uart_rx_data_callback(uint8_t uart_instance, uint8_t *dta, uint16_t length);
 /* Timer callback */
 void vTimerCallback(TimerHandle_t pxTimer);
-/* Convert char to dec */
-uint32_t app_convert_char2Dec(uint8_t ch1, uint8_t ch2);
 
 /* Main app function */
 void app_main(void)
@@ -229,6 +228,10 @@ void app_uart_rx_data_callback(uint8_t uart_instance, uint8_t *dta, uint16_t len
 
             app_parse_data_from_uart(dta, length);
         }
+        else
+        {
+            ESP_LOGI(TAG, "%s", dta);
+        }
     }
 }
 
@@ -303,38 +306,4 @@ static void app_parse_data_from_uart(uint8_t data[], uint32_t length)
             }
         }
     }
-}
-
-/* Convert char to dec */
-uint32_t app_convert_char2Dec(uint8_t ch1, uint8_t ch2)
-{
-    uint32_t retVal = 0;
-
-    if((ch1 >= '0') && (ch1 <= '9'))
-    {
-        retVal = ((ch1 - 48) * 16);
-    }
-    else if((ch1 >= 'A') && (ch1 <= 'F'))
-    {
-        retVal = ((ch1 - 55) * 16);
-    }
-    else if((ch1 >= 'a') && (ch1 <= 'f'))
-    {
-        retVal = ((ch1 - 87) * 16);
-    }
-
-    if((ch2 >= '0') && (ch2 <= '9'))
-    {
-        retVal += (ch2 - 48);
-    }
-    else if((ch2 >= 'A') && (ch2 <= 'F'))
-    {
-        retVal += (ch2 - 55);
-    }
-    else if((ch2 >= 'a') && (ch2 <= 'f'))
-    {
-        retVal += (ch2 - 87);
-    }
-
-    return retVal; 
 }
